javacode/meta-dbt/requirements.txt
pymysql==1.0.2
PyYAML==6.0
requests==2.28.1